seperate month slice
