+++ src/redux/reducers/dayReducer.ts
@@ -20,7 +20,6 @@ export interface weekInterface {
 
 
 export interface initialStateProp {
-    month: weekInterface[],
     activeWeekMoment?: dayInterface[],
     hoursOfDay: string[],
 }
@@ -38,35 +37,7 @@ for (let i = 0; i < 7; i++) {
 }
 
 
-// Get the current month and calculate the number of weeks in it
-let currentMonth = moment().startOf('month');
-// let weeksInMonth = currentMonth.isoWeeksInMonth();
-
-let monthWeeks = []
-
-// Create an array of weeks for the current month
-for (let i = 0; i < 6; i++) {
-    let currentWeekStart = currentMonth.clone().startOf('week');
-    let initalDays = []
-
-    for (let j = 0; j < 7; j++) {
-        initalDays.push({
-            day: currentWeekStart.format('dddd'),
-            date: currentWeekStart.clone(),
-            assignedHours: []
-        });
-        currentWeekStart.add(1, 'day');
-    }
-
-    monthWeeks.push({
-        daysOfWeek: initalDays,
-        currentWeekMoment: currentWeekStart.clone()
-    });
-    currentMonth.add(1, 'week');
-}
-
 const initialState: initialStateProp = {
-    month: monthWeeks,
     activeWeekMoment: initalDays,
     hoursOfDay: ['12am', '1am', '2am', '3am', '4am', '5am', '6am', '7am', '8am', '9am', '10am', '11am', '12pm', '1pm', '2pm', '3pm', '4pm', '5pm', '6pm', '7pm', '8pm', '9pm', '10pm', '11pm'],
 }
@@ -89,6 +60,9 @@ export const weekSlice = createSlice({
         setnewCurrentWeekMoment: (state, action) => {
             // state.currentWeekMoment = action.payload.clone()
         },
+        setActiveWeek: (state, action) => {
+            state.activeWeekMoment = action.payload
+        },
         addAssignedHour: (state, action) => {
             let dayIndex = action.payload.index;
             let hour = action.payload.time;
@@ -101,6 +75,6 @@ export const weekSlice = createSlice({
     },
 })
 
-export const { setCurrentWeek, setnewCurrentWeekMoment, addAssignedHour } = weekSlice.actions
+export const { setCurrentWeek, setActiveWeek, setnewCurrentWeekMoment, addAssignedHour } = weekSlice.actions
 
 export default weekSlice.reducer
